# src/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "MiniSense RESTful API Documentation",
        "contact": {
            "email": "alefs07@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3333/",
            "description": "MiniSense RESTful API test"
        }
    ],
    "paths": {
...
        "/user": {
            "post": {
                "summary": "User registration",
                "description": "Route responsible for registering a new user. No parameters are needed in the request but you must assign a body in json format with username and email. If the username or email of the body already exists in the database you will receive a 400 response, but if they do not exist you will receive a 200 response with a json with id, username and email.",
                "tags": ["User"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            },
                            "examples": {
                                "user": {
                                    "value": {
                                        "username": "my_username",
                                        "email": "my_email@gmail.com"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Username or email already exists."
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/User"
                                },
                                "examples": {
                                    "user": {
                                        "value": {
                                            "id": 1,
                                            "username": "my_username",
                                            "email": "my_email@hotmail.com"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "List of all registered users",
                "description": "Route responsible for listing all registered users. No parameters are needed in the request and no body. In response you will receive a 200 response with a json with an array with all registered users or an empty array.",
                "tags": ["User"],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "$ref": "#/components/schemas/User"
                                },
                                "examples": {
                                    "user": {
                                        "value": [
                                            {
                                                "id": 1,
                                                "username": "my_email",
                                                "email": "my_email@gmail.com"
                                            },
                                            {
                                                "id": 2,
                                                "username": "support",
                                                "email": "support@gmail.com"
                                            },
                                            {
                                                "id": 3,
                                                "username": "aux",
                                                "email": "aux@gmail.com"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{email}": {
            "get": {
                "summary": "Query a user by their email",
                "description": "Route responsible for consulting a user by their email. You must pass a user's email as a parameter. A body is not needed. In response, if the email does not exist in the database you will receive a 400 response, but if the email exists you will receive a 200 response with a json with: id, username and email.",
                "tags": ["User"],
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "User email for consultation",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "User not found."
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/User"
                                },
                                "examples": {
                                    "user": {
                                        "value": {
                                            "id": 1,
                                            "username": "my_email",
                                            "email": "my_email@hotmail.com"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
                
            },
...
        }
    }
}